Fixed conflicts:
	doc/userguide/models/fs2000ns_steadystate.m
	doc/userguide/models/fsdat.m
	matlab/modules/dseries
	mex/sources/estimation/tests/logposterior_dll_test/random_walk_metropolis_hastings_core.m
	mex/sources/estimation/tests/logposterior_dll_test/rawdata_euromodel_1.m
	mex/sources/estimation/tests/random_walk_metropolis_hastings_core.m
	mex/sources/estimation/tests/rawdata_euromodel_1.m

+++ preprocessor/SymbolTable.cc
@@ -110,7 +110,7 @@ SymbolTable::freeze() throw (FrozenException)
 
   frozen = true;
 
-  for (int i = 0; i < symbol_table.size(); i++)
+  for (int i = 0; i < (int) symbol_table.size(); i++)
     {
       int tsi;
       switch (getType(i))
